+++ animation_nodes/ui/node_menu.py
@@ -265,6 +265,7 @@ class ListMenu(bpy.types.Menu):
         layout.separator()
         insertNode(layout, "an_FillListNode", "Fill")
         insertNode(layout, "an_RepeatListNode", "Repeat")
+        insertNode(layout, "an_RepeatListElementsNode", "Repeat Elements")
         insertNode(layout, "an_RandomListNode", "Random")
         insertNode(layout, "an_ListBooleanOperationsNode", "List Boolean Operations")
         layout.separator()