+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added *Normals* output in *Mesh Points Scatter* node.
 - Added Interpolation input to *Remap Falloff* node.
 - Added *Decompose Text* node to menu.
+- Added *Repeat List Elements* node.
 
 ### Fixed